ai-therapist-fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
groq==0.15.0
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
tensorflow==2.15.0
keras==2.15.0
Pillow==10.1.0